Divorce et séparation aux USA (Rosenfeld & Roesler, 2024)

Les données sont reprises du "replication package" de Rosenfeld & Roesler (2024). Dans cet article, les auteur·ices mesurent l'évolution dans le temps des facteurs prédictifs de la dissolution des mariages aux États-Unis. Les données de la National Survey of Family Growth sont mobilisées, couvrant sept décennies d'histoires matrimoniales.

Les données se concentrent sur les premiers mariages des femmes. Elles sont particulièrement adaptées à l'analyse de durées en présence de censure à droite et de variables qui changent dans le temps.

Usage

data("divorce")

Format

Data frame avec 470232 observations et 16 variables.

case_id

(character) identifiant unique des enquêtées

divorce_separation1

(integer) 1 si divorce ou séparation durant l'année, 0 sinon

mar1_length

(integer) nombre d'années écoulées depuis le début du premier mariage

age_at_mar1

(factor) âge au premier mariage (less than 18, 18-19, 20-24, 25 and more)

any_children_u18

(factor) présence d'un enfant de moins de 18 ans (no, yes)

marriage1_cohort

(integer) cohorte du premier mariage

premar1

(factor) cohabitation pré-maritale (no, yes)

educ

(factor) niveau d'éducation (less than HS, HS degree, some college, BA or more)

race

(factor) race (black, white, other)

both_parents_together_14

(factor) parents en couple quand l'enquêtée avait 14 ans (no, yes)

mother_educ

(factor) niveau d'éducation de la mère (less than HS, HS, some college, BA or more)

survey_year

(integer) année de l'enquête

year

(integer) année d'observation

wave

(numeric) vague de l'enquête

children

(integer) nombre d'enfants

religion

(factor) religion (none, catholic, protestant, other)

Details

Les observations sont au format personnes-années. Il y a 49181 enquêtées distinctes (les données ne décrivent que des femmes).

Source

Les données et scripts R pour la réplication de l'article sont accessibles ici : https://web.stanford.edu/~mrosenfe/#my_papers

References

Rosenfeld Michael J. & Roesler Katharina, 2024, "Stability and change in predictors of marital dissolution in the US 1950–2017", Journal of Marriage and Family, 86(1), 154–175. https://doi.org/10.1111/jomf.12932