Activité, emploi et chômage en 2022

Enquête emploi en continu (Insee, 2022) - Fichier détail

Usage

data("emploi")

Format

Data frame avec 348964 observations et 73 variables

Details

L’enquête Emploi, déclinaison française des enquêtes "Forces de travail" ("Labour Force Survey") définies au niveau européen, est la seule source permettant de mesurer le chômage et l’activité au sens du Bureau international du travail (BIT). Elle est menée en continu, sur l’ensemble des semaines de l’année, en France hors Mayotte. En 2022, chaque trimestre, environ 90 000 personnes de 15 ans ou plus vivant en ménage ordinaire (c’est-à-dire hors foyers, hôpitaux, prisons...) ont répondu à l’enquête. Les personnes décrivent leur situation vis-à-vis du marché du travail au cours d’une semaine donnée, dite « de référence ». En moyenne annuelle, le taux de chômage est estimé avec une précision à 95

En 2021, l’enquête Emploi a connu une refonte majeure. Son questionnaire a été rénové afin, d’une part, de se conformer au nouveau règlement européen sur les statistiques sociales qui l’encadre et, d’autre part, de mieux répondre aux besoins de ses utilisateurs et aux nouveaux enjeux du marché du travail. Son protocole de collecte a également été modernisé (possibilité de répondre par Internet en réinterrogation) et sa méthode de pondération revue.

Source

L'ensemble des fichiers et de la documentation (questionnaires, dictionnaire des variables, etc.) est consultable à l'adresse suivante : https://www.insee.fr/fr/statistiques/7657353