Enquête "Étude des Relations Familiales et Intergénérationnelles" (ERFI)

L'Enquête sur les Relations Familiales et Intergénérationnelles (ERFI) a pour objectif de décrire les situations familiales dans toute leur diversité et, notamment, les relations entretenues avec ses parents, enfants, petits-enfants, beaux-enfants mais aussi les relations de couples. Afin de comprendre ces comportements, plusieurs milliers de personnes ont été suivies pendant plusieurs années.

Usage

data("erfi")

Format

A data frame with 10079 observations and 86 variables.

Le data frame contient des labels de variables et de modalités de variables, manipulables avec le package labelled (https://larmarange.github.io/labelled/articles/labelled.html)

Details

L'Enquête sur les Relations Familiales et Intergénérationnelles (ERFI) a pour objectif de décrire les situations familiales dans toute leur diversité et, notamment, les relations entretenues avec ses parents, enfants, petits-enfants, beaux-enfants mais aussi les relations de couples. Afin de comprendre ces comportements, plusieurs milliers de personnes ont été suivies pendant plusieurs années. Pour plus d'informations sur l'enquête, veuillez vous référez aux métadonnées complètes (https://doi.org/10.48756/ined-IE0214-4582). Le questionnaire et le dictionnaire des codes sont dans la rubrique "ressources téléchargeables".

Le fichier des données anonymisées de l'enquête ERFI (vague 1 - 2005) a été produit par le Service des Enquêtes et Sondages (SES) de l’Ined dans le cadre d’une formation organisée par les Plateformes Universitaires de Données (PUD) de Bordeaux et Strasbourg. Ce fichier permet de reproduire, aussi fidèlement que possible, les résultats d’une publication tirée du bulletin d’information scientifique de l’Ined, le Population & Sociétés n°427 (https://www.ined.fr/fichier/s_rubrique/19095/pop_et_soc_francais_427.fr.pdf). Il est un extrait anonymisé du Fichier de Production et de Recherche de l'enquête ERFI (vague 1 - 2005), mis à disposition par l'Institut national d'études démographiques (Ined) via la plateforme Quetelet-Progedo-Diffusion. Il comporte 86 variables permettant de reproduire certains des tableaux et statistiques présentés dans le numéro de Population & Sociétés cité ci-avant. Il permet de réaliser des statistiques simples sur la fréquence à laquelle on fréquente (ou non) ses parents.

Ce fichier de données est mis à disposition à des fins d'enseignement. Son utilisation ne peut donner lieu à des publications scientifiques.

Note

Ce jeu de données fait partie d'un kit d’initiation à l’exploitation de données d’enquête à partir du langage R, accessible ici : https://lifeobs.site.ined.fr/fr/outils-formation/kits-autoformation/initiation-r-erfi1/

Source

Toute réutilisation du fichier de données doit systématiquement faire mention de la citation suivante :

Données anonymisées - Enquête sur les relations familiales et intergénérationnelles (ERFI - 1ère vague) (2005) [fichier de données]. Ined [producteur]. Ined [diffuseur]. https://doi.org/10.48756/ined-IE0214AN-1

References

Régnier-Loilier, A. (2006). "À quelle fréquence voit-on ses parents ?", Population & Sociétés, 2006/9, n°427, pp. 1-4. https://www.ined.fr/fichier/s_rubrique/19095/pop_et_soc_francais_427.fr.pdf