Enquête "Le temps donné aux tableaux" (Passeron & Pedler, 1987)

Données de l'enquête emmenée par Jean-Claude Passeron et Emmanuel Pedler en 1987. Les enquêteurs ont suivi les visiteurs du musée Granet d'Aix-en-Provence, chronomètre en main, avec pour objectif de mettre en évidence les ressorts sociaux du « temps donné aux tableaux ».

Usage

data("tdt")

Format

Data frame avec 144 observations et 37 variables.

Source

Les données sont accessibles depuis les annexes du la réédition numérique de l'ouvrage tiré de l'enquête. On y trouve également le dictionnaire des variables.

https://books.openedition.org/enseditions/16817

References

Passeron Jean-Claude, Pedler Emmanuel Pedler, 2019, Le temps donné aux tableaux, édité par Samuel Coavoux, Igor Martinache, et Pierre Mercklé. Lyon: ENS Éditions, 2019. https://doi.org/10.4000/books.enseditions.16682.