Enquête "Trajectoires et Origines" (TeO-1)

L'enquête Trajectoires et Origines (TeO-1) a eu pour objectif d'appréhender dans quelle mesure l'origine nationale ou migratoire, combinée à d'autres facteurs communs à toute la société française (âge, genre, niveau d'instruction, catégorie sociale, lieu de résidence, etc.), contribue à expliquer les inégalités d'accès au logement, à l'emploi, aux loisirs, aux services publics et à d'autres ressources de la vie sociale.

Usage

data("teo1")

Format

A data frame with 21761 observations on the following 53 variables.

Le data frame contient des labels de variables et de modalités de variables, manipulables avec le package labelled (https://larmarange.github.io/labelled/articles/labelled.html)

Details

L'enquête Trajectoires et Origines (TeO) a eu pour objectif d'appréhender dans quelle mesure l'origine nationale ou migratoire, combinée à d'autres facteurs communs à toute la société française (âge, genre, niveau d'instruction, catégorie sociale, lieu de résidence, etc.), contribue à expliquer les inégalités d'accès au logement, à l'emploi, aux loisirs, aux services publics et à d'autres ressources de la vie sociale. Pour plus d'informations sur l'enquête, veuillez vous référez aux métadonnées complètes (https://doi.org/10.13144/lil-0494). Le questionnaire et le dictionnaire des codes sont dans la rubrique "ressources téléchargeables".

Le fichier des données anonymisées de l'enquête TeO (2008) a été produit par le Service des Enquêtes et Sondages (SES) de l’Ined dans le cadre d’une formation organisée par le Service des Méthodes Statistiques (SMS) de l’Ined et la Plateforme Universitaire de Données des Grands-Moulins. Ce fichier permet de reproduire, aussi fidèlement que possible, les résultats d’une publication tirée du bulletin d’information scientifique de l’Ined, le Population & Sociétés n°466 (https://doi.org/10.3917/popsoc.466.0001). Il est un extrait anonymisé du Fichier de Production et de Recherche de l'enquête TeO (2008), mis à disposition par l'Archives des Données de la Statistique Publique (ADISP) via la plateforme Quetelet-Progedo-Diffusion. Il comporte 53 variables permettant de reproduire certains des tableaux et statistiques présentés dans le numéro de Population & Sociétés cité ci-avant. Il permet de réaliser des statistiques simples sur la perception de la discrimination et des raisons de cette discrimination.

Ce fichier de données est mis à disposition à des fins d'enseignement. Son utilisation ne peut donner lieu à des publications scientifiques.

Note

Ce jeu de données fait partie d'un kit d’initiation à l’exploitation de données d’enquête à partir du langage R, accessible ici : https://lifeobs.site.ined.fr/fr/outils-formation/supports-de-formation/initiation-a-exploitation-de-donnees-denquete-a-partir-du-langage-r-teo-1/

Source

Toute réutilisation du fichier de données doit systématiquement faire mention de la citation suivante :

Données anonymisées - Enquête Trajectoires et Origines (TeO) (2008) [fichier de données]. Ined [producteur]. Ined [diffuseur]. https://doi.org/10.48756/ined-IE0243AN-1

References

Beauchemin Cris, Hamel Christelle, Lesné Maud et al. "Les discriminations: une question de minorités visibles", Population & Sociétés, 2010, n°4, pp.1-4. https://doi.org/10.3917/popsoc.466.0001